api/requirements.txt
opencv-python
numpy
tensorflow == 2.12.0
django
djangorestframework
django-cors-headers